+++ src/components/Dropdown.js
@@ -180,7 +180,7 @@ Dropdown.propTypes = {
   displayKey: PropTypes.string.isRequired,
   dark: PropTypes.bool,
   monospace: PropTypes.bool,
-  selected: PropTypes.string,
+  selected: PropTypes.any,
   onChange: PropTypes.func,
   iconColor: PropTypes.string,
 };