+++ src/handlers/ledger-eth.js
@@ -2,10 +2,7 @@ import EthereumTx from 'ethereumjs-tx';
 import TransportU2F from '@ledgerhq/hw-transport-u2f';
 import AppEth from '@ledgerhq/hw-app-eth';
 import ethereumNetworks from '../references/ethereum-networks.json';
-import {
-  removeHexPrefix,
-  getDerivationPathComponents,
-} from '../helpers/utilities';
+import { removeHexPrefix, getDerivationPathComponents } from 'balance-common';
 
 /**
  * @desc Ledger ETH App instance