+++ src/layouts/base.js
@@ -2,7 +2,7 @@ import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
-import lang from '../languages';
+import { lang, resources } from 'balance-common';
 import Link from '../components/Link';
 import Dropdown from '../components/Dropdown';
 import Background from '../components/Background';
@@ -23,15 +23,14 @@ import {
   accountClearState,
   accountUpdateAccountAddress,
   accountChangeLanguage,
-} from '../reducers/_account';
+} from 'balance-common';
 import { metamaskClearState } from '../reducers/_metamask';
 import { ledgerClearState } from '../reducers/_ledger';
 import { trezorClearState } from '../reducers/_trezor';
 import { walletConnectClearState } from '../reducers/_walletconnect';
-import { resetAccount, resetWalletConnect } from '../handlers/localstorage';
+import { commonStorage } from 'balance-common';
 import ReminderRibbon from '../components/ReminderRibbon';
 import { colors, responsive } from '../styles';
-import { resources } from '../languages';
 
 const StyledLayout = styled.div`
   position: relative;
@@ -180,14 +179,14 @@ const BaseLayout = ({
   const openSendModal = () => modalOpen('DONATION_MODAL');
   const disconnectAccount = () => {
     accountClearState();
-    resetAccount(accountAddress);
+    commonStorage.resetAccount(accountAddress);
     if (accountType === 'TREZOR') {
       trezorClearState();
     } else if (accountType === 'LEDGER') {
       ledgerClearState();
     } else if (accountType === 'WALLETCONNECT') {
       walletConnectClearState();
-      resetWalletConnect();
+      commonStorage.resetWalletConnect();
     } else if (accountType === 'METAMASK') {
       metamaskClearState();
     }