+++ src/pages/Wallet.js
@@ -6,7 +6,7 @@ import BaseLayout from '../layouts/base';
 import Account from '../views/Account';
 import Card from '../components/Card';
 import { walletConnectHasValidSession } from '../reducers/_walletconnect';
-import lang from '../languages';
+import { lang } from 'balance-common';
 import { fonts, colors } from '../styles';
 
 const StyledWrapper = styled.div`