+++ src/views/Account/AccountBalances.js
@@ -2,15 +2,15 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import lang from '../../languages';
 import AssetIcon from '../../components/AssetIcon';
 import ToggleIndicator from '../../components/ToggleIndicator';
-import { ellipseText } from '../../helpers/utilities';
 import {
   convertStringToNumber,
+  ellipseText,
   hasHighMarketValue,
   hasLowMarketValue,
-} from '../../helpers/bignumber';
+  lang,
+} from 'balance-common';
 import { colors, fonts, responsive } from '../../styles';
 
 const StyledGrid = styled.div`
@@ -313,7 +313,7 @@ class AccountBalances extends Component {
           ) : (
             <div />
           )}
-          <p>{`${accountInfo.total.display || '———'}`}</p>
+          <p>{`${accountInfo.total ? accountInfo.total.display : '———'}`}</p>
         </StyledLastRow>
       </StyledGrid>
     );